:snowflake: Change way to check for "pip".
When a push were performed to deploy a project, the local "pip"
and "Python" were the target to run the deploy commands.

+++ utils/check_dependencies.sh
@@ -11,13 +11,15 @@ ORIGINAL_PATH=`pwd` # Save current directory
 source ${UTILS_PATH}/log_messages.sh
 
 # Install/Upgrade pip
-if [ ! $(command -v pip) > /dev/null ] || [[ ! $(pip --version) =~ $(echo "^pip (9|1[0-9])\.[0-9]+\.[1-9]") ]]; then
-  log "Install/upgrade \"pip\""
-  cd ${TMP_PATH}
-  curl -OL https://bootstrap.pypa.io/get-pip.py || exit 1
-  python2.7 get-pip.py
-  rm -rf get-pip.py
-  cd ${ORIGINAL_PATH}
+if [ ${CHECK_PIP} ]; then
+  if [ ! $(command -v pip) > /dev/null ] || [[ ! $(pip --version) =~ $(echo "^pip (9|1[0-9])\.[0-9]+\.[1-9]") ]]; then
+    log "Install/upgrade \"pip\""
+    cd ${TMP_PATH}
+    curl -OL https://bootstrap.pypa.io/get-pip.py || exit 1
+    python2.7 get-pip.py
+    rm -rf get-pip.py
+    cd ${ORIGINAL_PATH}
+  fi
 fi
 
 # Check for JQ https://stedolan.github.io/jq/
